src/testbed/ml/pyproject.toml
[project]
name = "ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.10.3",
    "pandas>=2.3.1",
    "python-socketio>=5.13.0",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "websocket-client>=1.8.0",
]
